Increase view number without modulo to allow easier comparisons
The view change & recovery protocol need to understand easily if a
view is "larger" than another one. This might be tricky if we use
cyclic view number. Thus, we will have a infinite counter for the view
number and use the modulo to check if the replica is a primary.