Replace `Network` by `IO`
This is the main commit that replaces our own implementation of the
`Network` between replicas with an internal message bus using a
custom `IO` implementation based on `mio`.

A couple of things are still missing, such as:
 * Correct error handling
 * Simplify bus for replica vs client
 * Fix unit test setup

But this already works for the "simple" example, and we can further
iterate on this.

This should allow us to implement simulation testing by mocking the
IO implementation with probabilistic models.